Move CategoryDropdown to its own folder
